Extract shared DOM assertion macros
